SurveyDefsLoader.js: moved to webapp/components/survey, added index.js and updated imports
